Cleanup

+++ src/bsp_util.cc
@@ -140,7 +140,7 @@ angle_g UtilComputeAngle(double dx, double dy)
 	if (dx == 0)
 		return (dy > 0) ? 90.0 : 270.0;
 
-	angle = atan2((double) dy, (double) dx) * 180.0 / M_PI;
+	angle = atan2(dy, dx) * 180.0 / M_PI;
 
 	if (angle < 0)
 		angle += 360.0;
